+++ cmd/start.go
@@ -79,29 +79,23 @@ var startCmd = &cobra.Command{
 		// create new k8s controller
 		ctlr, err := controller.New()
 		if err != nil {
-			logrus.Panic(err.Error())
+			logrus.Fatalf("could not create controller: %s", err.Error())
 			os.Exit(1)
 		}
 
 		// load decryption key into memory
 		if err := utils.LoadDecryptionKey(viper.GetString("decryption-key-file")); err != nil {
-			logrus.Panic(err.Error())
+			logrus.Fatalf("could not load decryption key: %s", err.Error())
 			os.Exit(1)
 		}
 
 		// create tprs
 		if err := ctlr.CreateTPRs(); err != nil {
-			logrus.Panic(err.Error())
+			logrus.Fatalf("could not create TPRs: %s", err.Error())
 			os.Exit(1)
 		}
 
-		// start walking k8s resources
-		go func() {
-			if err := ctlr.Watch(); err != nil {
-				logrus.Fatal(err.Error())
-				os.Exit(1)
-			}
-		}()
+		go ctlr.Watch()
 
 		// start UDP server
 		go func() {
@@ -111,38 +105,29 @@ var startCmd = &cobra.Command{
 			}
 		}()
 
-		// potentially start tracing server
-		if viper.GetBool("enable-tracing") {
-			tracer, closer, err := tracer.Jaeger()
-			if err != nil {
-				logrus.Fatal(err.Error())
-				os.Exit(1)
-			}
-			logrus.Infof("starting global tracer")
+		tracer, closer, err := tracer.Jaeger()
+		if err != nil {
+			logrus.Warnf("error create Jaeger tracer: %s", err.Error())
+		} else {
 			opentracing.SetGlobalTracer(tracer)
 			defer func() {
 				if err := closer.Close(); err != nil {
-					logrus.Warnf("there was a problem closing the tracer: %s", err.Error())
+					logrus.Warnf("error closing Jaeger tracer: %s", err.Error())
 				}
 			}()
 		}
 
-		// attempt to create influxdb client
 		influxCtlr, err := monitor.NewInfluxdbController()
 		if err != nil {
-			logrus.Warnf("there was an error connecting to influxdb - analytics and monitoring will not be available", err.Error())
+			logrus.Warnf("error connecting to InfluxDB: %s", err.Error())
 		} else {
 			defer func() {
 				if err := influxCtlr.Client.Close(); err != nil {
-					logrus.Warnf("there was a problem closing the connection to influxdb: %s", err.Error())
+					logrus.Warnf("error closing the connection to InfluxDB: %s", err.Error())
 				}
 			}()
 		}
 
-		// start kanali readiness server
-		go server.StatusServer(ctlr)
-
-		// start kanali gateway server
 		server.Start(ctlr, influxCtlr)
 
 	},